CAMS-37: Use separate proxy [skip ci]
