+++ backend/functions/lib/adapters/data-access.proxy.ts
@@ -8,7 +8,6 @@ import config from '../configs/index';
 import { PersistenceGateway } from './types/persistence-gateway';
 import log from './services/logger.service';
 import { Context } from './types/basic';
-import {PacerGatewayInterface} from "../use-cases/pacer.gateway.interface";
 
 const NAMESPACE = 'DATA-ACCESS-PROXY';
 
@@ -33,15 +32,4 @@ async function proxyData(context: Context, table: string, mock: boolean = false)
   }
 }
 
-async function proxyPacer(context: Context): Promise<PacerGatewayInterface> {
-  if (config.get('pacerMock')) {
-    log.info(context, NAMESPACE, 'using local PACER data');
-    return await import(`./gateways/local.pacer.gateway`);
-  } else {
-    log.info(context, NAMESPACE, 'using PACER API');
-    return await import(`./gateways/pacer.gateway`);
-  }
-}
-
 export default proxyData;
-export { proxyPacer };
